Move CatalogueHeader redux down
